refactor(playlists): Introduce PlaylistsContext provider (#24)
In order to simplify the playlists fetching and remove the duplicated
code across all compontents, this change creates a shared state where
all components depends on for redering.

I found few problems with Spotify API where the changes take over 20
seconds to take effect. So, the songs counts on sidebar is not showing
the count anymore (removed due to the delay).  The name changing also
has a bit delay to take effect.